feat: refactor HomePage to use server components and improve data fetching for user-specific courses and roles
